tests/wpt/web-platform-tests/css/css-fonts/parsing/font-variant-numeric-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: getComputedStyle().fontVariantNumeric</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-numeric-prop">
<meta name="assert" content="font-variant-numeric computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-variant-numeric', 'normal');

// <numeric-figure-values>   = [ lining-nums | oldstyle-nums ]
test_computed_value('font-variant-numeric', 'lining-nums');
test_computed_value('font-variant-numeric', 'oldstyle-nums');

// <numeric-spacing-values>  = [ proportional-nums | tabular-nums ]
test_computed_value('font-variant-numeric', 'proportional-nums');
test_computed_value('font-variant-numeric', 'tabular-nums');

// <numeric-fraction-values> = [ diagonal-fractions | stacked-fractions ]
test_computed_value('font-variant-numeric', 'diagonal-fractions');
test_computed_value('font-variant-numeric', 'stacked-fractions');

test_computed_value('font-variant-numeric', 'ordinal');

test_computed_value('font-variant-numeric', 'slashed-zero');

test_computed_value('font-variant-numeric', 'oldstyle-nums tabular-nums diagonal-fractions');

test_computed_value('font-variant-numeric', 'lining-nums proportional-nums stacked-fractions ordinal slashed-zero');
</script>
</body>
</html>
